mmc: sdhci-msm: Correct clock doubling logic during HS400 mode transition
According to the hardware programming guide, the clock frequency must
remain below 52MHz during the transition to HS400 mode.

However,in the current implementation, the timing is set to HS400 (a
DDR mode) before adjusting the clock. This causes the clock to double
prematurely to 104MHz during the transition phase, violating the
specification and potentially resulting in CRC errors or CMD timeouts.

This change ensures that clock doubling is avoided during intermediate
transitions and is applied only when the card requires a 200MHz clock
for HS400 operation.

Change-Id: Ia6db0955549bc7aab8fa0ef55ce255e7f3dacc70
Signed-off-by: Sarthak Garg <quic_sartgarg@quicinc.com>