Revert "net: qrtr: mhi: Abort pending tx on device removal"
This reverts commit d13552a218da42e1b474383ae9679e2480461007.

Reason for revert: A new patchset update for the original
fix was missed during propagation. (patchset 6 is the final
version, but patchset 4 got merged, which didn't have
READ_ONCE and WRITE_ONCE changes)
Unblock any pending tx when device remove is called, so that
it won't get blocked in the wait for ringbuffer.

Change-Id: Id783e376405a78abb7458ecc0ec4a8edd26204d0
Signed-off-by: Sarannya S <sarannya.s@oss.qualcomm.com>